# swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "API Logic Server",
        "description": "SAFRSAPI",
        "termsOfService": "",
        "version": "0.0"
    },
    "host": "localhost:5656",
    "basePath": "/api",
    "paths": {
...
        "/Branch/": {
            "get": {
                "tags": [
                    "Branch"
                ],
                "summary": "Retrieve a collection of Branch objects",
                "parameters": [
                    {
                        "default": "CustomerList,EmployeeList",
                        "type": "string",
                        "name": "include",
                        "in": "query",
                        "format": "string",
                        "required": false,
                        "description": "Branch relationships to include (csv)"
                    },
                    {
                        "default": "BranchID,Name,Office,Address,OpenDate,_check_sum_,S_CheckSum",
                        "type": "string",
                        "name": "fields[Branch]",
                        "in": "query",
                        "format": "string",
                        "required": false,
                        "description": "Branch fields to include (csv)"
                    },
                    {
                        "default": 0,
                        "type": "integer",
                        "name": "page[offset]",
                        "in": "query",
                        "format": "int64",
                        "required": false,
                        "description": "Page offset"
                    },
                    {
                        "default": 10,
                        "type": "integer",
                        "name": "page[limit]",
                        "in": "query",
                        "format": "int64",
                        "required": false,
                        "description": "Max number of items"
                    },
                    {
                        "default": "id",
                        "type": "string",
                        "name": "sort",
                        "in": "query",
                        "format": "string",
                        "required": false,
                        "description": "Sort order"
                    },
                    {
                        "default": "",
                        "type": "string",
                        "name": "filter[BranchID]",
                        "in": "query",
                        "format": "string",
                        "required": false,
                        "description": "BranchID attribute filter (csv)"
                    },
                    {
                        "default": "",
                        "type": "string",
                        "name": "filter[Name]",
                        "in": "query",
                        "format": "string",
                        "required": false,
                        "description": "Name attribute filter (csv)"
                    },
                    {
                        "default": "",
                        "type": "string",
                        "name": "filter[Office]",
                        "in": "query",
                        "format": "string",
                        "required": false,
                        "description": "Office attribute filter (csv)"
                    },
                    {
                        "default": "",
                        "type": "string",
                        "name": "filter[Address]",
                        "in": "query",
                        "format": "string",
                        "required": false,
                        "description": "Address attribute filter (csv)"
                    },
                    {
                        "default": "",
                        "type": "string",
                        "name": "filter[OpenDate]",
                        "in": "query",
                        "format": "string",
                        "required": false,
                        "description": "OpenDate attribute filter (csv)"
                    },
                    {
                        "default": "",
                        "type": "string",
                        "name": "filter[_check_sum_]",
                        "in": "query",
                        "format": "string",
                        "required": false,
                        "description": "_check_sum_ attribute filter (csv)"
                    },
                    {
                        "default": "",
                        "type": "string",
                        "name": "filter[S_CheckSum]",
                        "in": "query",
                        "format": "string",
                        "required": false,
                        "description": "S_CheckSum attribute filter (csv)"
                    },
                    {
                        "default": "",
                        "type": "string",
                        "name": "filter[id]",
                        "in": "query",
                        "format": "string",
                        "required": false,
                        "description": "id attribute filter (csv)"
                    },
                    {
                        "default": "",
                        "type": "string",
                        "name": "filter",
                        "in": "query",
                        "format": "string",
                        "required": false,
                        "description": "Custom Branch filter"
                    },
                    {
                        "name": "Content-Type",
                        "in": "header",
                        "type": "string",
                        "default": "application/vnd.api+json",
                        "enum": [
                            "application/vnd.api+json",
                            "application/json"
                        ],
                        "required": true
                    }
                ],
                "responses": {
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/jsonapi_error_403"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/jsonapi_error_404"
                        }
                    },
                    "200": {
                        "schema": {
                            "$ref": "#/definitions/Branch_coll1"
                        },
                        "description": "Request fulfilled, document follows"
                    }
                },
                "produces": [
                    "application/vnd.api+json"
                ],
                "description": "Retrieve Branch from Branch",
                "operationId": "RetrieveacollectionofBranchobjects_0"
            },
            "post": {
                "tags": [
                    "Branch"
                ],
                "summary": "Create Branch",
                "parameters": [
                    {
                        "name": "Content-Type",
                        "in": "header",
                        "type": "string",
                        "default": "application/vnd.api+json",
                        "enum": [
                            "application/vnd.api+json",
                            "application/json"
                        ],
                        "required": true
                    },
                    {
                        "name": "POST body",
                        "in": "body",
                        "description": "Branch attributes",
                        "schema": {
                            "$ref": "#/definitions/Branch_inst1"
                        },
                        "required": true
                    }
                ],
                "responses": {
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/jsonapi_error_403"
                        }
                    },
                    "202": {
                        "description": "Accepted"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/jsonapi_error_404"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/jsonapi_error_409"
                        }
                    },
                    "201": {
                        "schema": {
                            "$ref": "#/definitions/Branch_inst1"
                        },
                        "description": "Created"
                    }
                },
                "produces": [
                    "application/vnd.api+json"
                ],
                "operationId": "CreateBranch_0"
            }
        },
        "/Branch/{BranchId}/": {
            "get": {
                "tags": [
                    "Branch"
                ],
                "summary": "Retrieve Branch instance",
                "parameters": [
                    {
                        "default": "CustomerList,EmployeeList",
                        "type": "string",
                        "name": "include",
                        "in": "query",
                        "format": "string",
                        "required": false,
                        "description": "Branch relationships to include (csv)"
                    },
                    {
                        "default": "BranchID,Name,Office,Address,OpenDate,_check_sum_,S_CheckSum",
                        "type": "string",
                        "name": "fields[Branch]",
                        "in": "query",
                        "format": "string",
                        "required": false,
                        "description": "Branch fields to include (csv)"
                    },
                    {
                        "name": "BranchId",
                        "in": "path",
                        "type": "string",
                        "default": "1",
                        "required": true
                    },
                    {
                        "name": "Content-Type",
                        "in": "header",
                        "type": "string",
                        "default": "application/vnd.api+json",
                        "enum": [
                            "application/vnd.api+json",
                            "application/json"
                        ],
                        "required": true
                    }
                ],
                "responses": {
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/jsonapi_error_403"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/jsonapi_error_404"
                        }
                    },
                    "200": {
                        "schema": {
                            "$ref": "#/definitions/Branch_inst1"
                        },
                        "description": "Request fulfilled, document follows"
                    }
                },
                "produces": [
                    "application/vnd.api+json"
                ],
                "description": "Retrieve Branch from Branch",
                "operationId": "RetrieveBranchinstance_0"
            },
            "patch": {
                "tags": [
                    "Branch"
                ],
                "parameters": [
                    {
                        "name": "BranchId",
                        "in": "path",
                        "type": "string",
                        "default": "1",
                        "required": true
                    },
                    {
                        "name": "Content-Type",
                        "in": "header",
                        "type": "string",
                        "default": "application/vnd.api+json",
                        "enum": [
                            "application/vnd.api+json",
                            "application/json"
                        ],
                        "required": true
                    },
                    {
                        "name": "PATCH body",
                        "in": "body",
                        "description": "Branch attributes",
                        "schema": {
                            "$ref": "#/definitions/Branch_inst"
                        },
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted"
                    },
                    "204": {
                        "description": "No Content"
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/jsonapi_error_403"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/jsonapi_error_409"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/jsonapi_error_404"
                        }
                    },
                    "200": {
                        "schema": {
                            "$ref": "#/definitions/Branch_inst1"
                        },
                        "description": "Request fulfilled, document follows"
                    }
                },
                "produces": [
                    "application/vnd.api+json"
                ],
                "summary": "Update Branch",
                "description": "Update Branch attributes",
                "operationId": "UpdateBranch_0"
            },
            "delete": {
                "tags": [
                    "Branch"
                ],
                "parameters": [
                    {
                        "name": "BranchId",
                        "in": "path",
                        "type": "string",
                        "default": "1",
                        "required": true
                    },
                    {
                        "name": "Content-Type",
                        "in": "header",
                        "type": "string",
                        "default": "application/vnd.api+json",
                        "enum": [
                            "application/vnd.api+json",
                            "application/json"
                        ],
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted"
                    },
                    "204": {
                        "description": "Request fulfilled, nothing follows"
                    },
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/Branch_inst1"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/jsonapi_error_403"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/jsonapi_error_404"
                        }
                    }
                },
                "produces": [
                    "application/vnd.api+json"
                ],
                "summary": "Delete Branch from Branch",
                "operationId": "DeleteBranchfromBranch_0"
            }
        },
...
    },
    "definitions": {
...
        "jsonapi_error_403": {
            "properties": {
                "errors": {
                    "example": [
                        {
                            "title": "Request forbidden -- authorization will not help",
                            "detail": "",
                            "code": "403"
                        }
                    ],
                    "type": "string"
                }
            },
            "description": ""
        },
        "jsonapi_error_404": {
            "properties": {
                "errors": {
                    "example": [
                        {
                            "title": "Nothing matches the given URI",
                            "detail": "",
                            "code": "404"
                        }
                    ],
                    "type": "string"
                }
            },
            "description": ""
        },
...
        "jsonapi_error_409": {
            "properties": {
                "errors": {
                    "example": [
                        {
                            "title": "Request conflict",
                            "detail": "",
                            "code": "409"
                        }
                    ],
                    "type": "string"
                }
            },
            "description": ""
        },
...
        "Branch_coll1": {
            "properties": {
                "data": {
                    "example": [
                        {
                            "attributes": {
                                "BranchID": 0,
                                "Name": "",
                                "Office": "",
                                "Address": "",
                                "OpenDate": "0001-01-01 00:00:00",
                                "_check_sum_": "",
                                "S_CheckSum": ""
                            },
                            "type": "Branch",
                            "id": "1",
                            "relationships": {
                                "CustomerList": {
                                    "data": [],
                                    "links": {
                                        "self": null
                                    }
                                },
                                "EmployeeList": {
                                    "data": [],
                                    "links": {
                                        "self": null
                                    }
                                }
                            }
                        }
                    ],
                    "type": "string"
                }
            },
            "description": "Branch get;"
        },
        "Branch_inst1": {
            "properties": {
                "data": {
                    "example": {
                        "attributes": {
                            "BranchID": 0,
                            "Name": "",
                            "Office": "",
                            "Address": "",
                            "OpenDate": "0001-01-01 00:00:00",
                            "_check_sum_": "",
                            "S_CheckSum": ""
                        },
                        "type": "Branch"
                    },
                    "type": "string"
                }
            },
            "description": "Branch post;"
        },
        "Branch_inst": {
            "properties": {
                "data": {
                    "example": {
                        "attributes": {
                            "BranchID": 0,
                            "Name": "",
                            "Office": "",
                            "Address": "",
                            "OpenDate": "0001-01-01 00:00:00",
                            "_check_sum_": "",
                            "S_CheckSum": ""
                        },
                        "type": "Branch",
                        "id": "1"
                    },
                    "type": "string"
                }
            },
            "description": "Branch patch;Branch post;Branch delete;Branch patch;Branch post;Branch delete;"
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "in": "header",
            "name": "Authorization"
        }
    },
    "security": [
        {
            "Bearer": []
        }
    ],
...
}